requirements.txt
langchain
langchain-community
langchain-openai
langchain-chroma
chromadb
pydantic
python-dotenv
fastapi
uvicorn
